eoogen@DESKTOP-00RM91I MINGW64 /d/nadynnnnnn/1840961-kekstagram-25 (module4-task1) $ git checkout master Switched to branch master Your branch is up to date with origin/master.
eoogen@DESKTOP-00RM91I MINGW64 /d/nadynnnnnn/1840961-kekstagram-25 (master)
$ git pu
pull   push

eoogen@DESKTOP-00RM91I MINGW64 /d/nadynnnnnn/1840961-kekstagram-25 (master)
$ git pu
pull   push

eoogen@DESKTOP-00RM91I MINGW64 /d/nadynnnnnn/1840961-kekstagram-25 (master)
$ git pull academy master
remote: Enumerating objects: 7, done.
remote: Counting objects: 100% (7/7), done.
remote: Compressing objects: 100% (3/3), done.
remote: Total 4 (delta 1), reused 1 (delta 0), pack-reused 0
Unpacking objects: 100% (4/4), done.
From github.com:htmlacademy-javascript/1840961-kekstagram-25
 * branch            master     -> FETCH_HEAD
   04940e2..0c0d5ac  master     -> academy/master
Updating 04c9c30..0c0d5ac
Fast-forward
 index.html |  2 +-
 js/main.js | 94 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 2 files changed, 95 insertions(+), 1 deletion(-)

eoogen@DESKTOP-00RM91I MINGW64 /d/nadynnnnnn/1840961-kekstagram-25 (master)
$ git push origin maste
error: src refspec maste does not match any.
error: failed to push some refs to git@github.com:Nadynn/1840961-kekstagram-25.git

eoogen@DESKTOP-00RM91I MINGW64 /d/nadynnnnnn/1840961-kekstagram-25 (master)
$ git push origin master
Total 0 (delta 0), reused 0 (delta 0)
To github.com:Nadynn/1840961-kekstagram-25.git
   04c9c30..0c0d5ac  master -> master

eoogen@DESKTOP-00RM91I MINGW64 /d/nadynnnnnn/1840961-kekstagram-25 (master)
$ git checkout -b module5-task1
Switched to a new branch module5-task1

eoogen@DESKTOP-00RM91I MINGW64 /d/nadynnnnnn/1840961-kekstagram-25 (module5-task1)
$ git status
On branch module5-task1
Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git checkout -- <file>..." to discard changes in working directory)

        modified:   index.html
        modified:   js/main.js

Untracked files:
  (use "git add <file>..." to include in what will be committed)

        js/data.js
        js/util.js

no changes added to commit (use "git add" and/or "git commit -a")

eoogen@DESKTOP-00RM91I MINGW64 /d/nadynnnnnn/1840961-kekstagram-25 (module5-task1)
$ git add *
The following paths are ignored by one of your .gitignore files:
node_modules
Use -f if you really want to add them.

eoogen@DESKTOP-00RM91I MINGW64 /d/nadynnnnnn/1840961-kekstagram-25 (module5-task1)
$ git status
On branch module5-task1
Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

        modified:   index.html
        new file:   js/data.js
        modified:   js/main.js
        new file:   js/util.js

eoogen@DESKTOP-00RM91I MINGW64 /d/nadynnnnnn/1840961-kekstagram-25 (module5-task1)
$ git commit -m
